db: Change cock_size to be a floating-point type

+++ db.py
@@ -10,7 +10,7 @@ class CockState(Base):
     __tablename__ = 'cock_state'
     chat_id: Mapped[int] = mapped_column(primary_key=True)
     user_id: Mapped[int] = mapped_column(primary_key=True)
-    cock_size: Mapped[int] = mapped_column(nullable=True)
+    cock_size: Mapped[float] = mapped_column(nullable=True)
     is_participating: Mapped[bool]
     active_event: Mapped[str] = mapped_column(nullable=True)
     event_duration: Mapped[int] = mapped_column(default=0)
